+++ @WVTransformHydrostatic/WVTransformHydrostatic.m
@@ -209,6 +209,9 @@ classdef WVTransformHydrostatic < WVGeometryDoublyPeriodicStratified & WVTransfo
             end
             self.shouldUseTrueNoMotionProfile = value;
             self.removeFromVariableCache("rho_nm");
+            self.removeFromVariableCache("eta_true");
+            self.removeFromVariableCache("ape");
+            self.removeFromVariableCache("apv");
         end
 
         function wvtX2 = waveVortexTransformWithResolution(self,m,options)